# backend/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Ghost Load API
  version: 1.0.0
  description: |
    API backend del MVP Ghost Load.

    Cubre:
    - Registro y continuidad de la evaluación del operador.
    - Calculadora de capacidad.
    - Benchmark de madurez.
    - Resultados consolidados.
    - Generación automática y descarga del PDF.
    - Importación de contactos.
    - Campañas de outreach.
    - Tracking de invitaciones.
    - Dashboard administrativo.

    Convenciones principales:
    - El correo identifica al operador.
    - El header `X-Evaluation-Token` autoriza el acceso a una evaluación.
    - El acceso administrativo utiliza JWT.
    - El PDF se genera automáticamente al completar el benchmark.
    - Los percentiles del MVP utilizan datos de demostración.
  contact:
    name: Ghost Load Team
servers:
  - url: http://localhost:8080
    description: Desarrollo local
  - url: https://api-staging.ghostload.example.com
    description: Staging
...
paths:
...
  /api/v1/admin/campaigns:
    post:
      tags: [Admin Campaigns]
      summary: Crear campaña
      operationId: createCampaign
      description: |
        Crea una campaña a partir de los contactos válidos de una importación
        completada y genera una invitación con token único por destinatario.
        Esta operación no envía correos.
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
      responses:
        '201':
          description: Campaña creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Admin Campaigns]
      summary: Listar campañas
      operationId: listCampaigns
      security:
        - adminBearer: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CampaignStatus'
      responses:
        '200':
          description: Campañas ordenadas desde la más reciente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CampaignSummaryResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/admin/campaigns/{campaignId}/send:
    parameters:
      - $ref: '#/components/parameters/CampaignId'
    post:
      tags: [Admin Campaigns]
      summary: Enviar o encolar invitaciones de una campaña
      description: >
        Cambia una campaña READY a SENDING y guarda sus correos en una cola
        durable de PostgreSQL. Un worker asíncrono realiza los envíos por SMTP.
      operationId: sendCampaign
      security:
        - adminBearer: []
      responses:
        '202':
          description: Envío aceptado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Campaña no enviable o ya procesada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    evaluationToken:
      type: apiKey
      in: header
      name: X-Evaluation-Token
      description: Token opaco generado al crear la evaluación.
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CampaignId:
      name: campaignId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Datos inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-07-23T20:10:00Z'
            status: 400
            code: VALIDATION_ERROR
            message: Uno o más campos son inválidos.
            path: /api/v1/evaluations/2f565fd3-f456-42e8-b9cb-552647d00655/calculator
            traceId: 7f9f827a2301
            fieldErrors:
              - field: productiveCapacityMw
                message: No puede superar totalCapacityMw.
    Unauthorized:
      description: Credencial ausente, inválida o expirada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Error interno no esperado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    CampaignStatus:
      type: string
      enum: [DRAFT, READY, SENDING, ACTIVE, COMPLETED, FAILED]
...
    CreateCampaignRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - subject
        - message
        - callToActionText
        - contactImportId
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 160
        description:
          type: [string, 'null']
          maxLength: 500
        subject:
          type: string
          minLength: 3
          maxLength: 180
        message:
          type: string
          minLength: 10
          maxLength: 5000
        callToActionText:
          type: string
          minLength: 2
          maxLength: 80
          example: Comenzar evaluación
        contactImportId:
          type: string
          format: uuid
        scheduledAt:
          type: [string, 'null']
          format: date-time
        timezone:
          type: [string, 'null']
          description: Identificador IANA de zona horaria.
          example: America/Lima
    CampaignResponse:
      type: object
      required:
        - id
        - name
        - status
        - subject
        - recipientCount
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          $ref: '#/components/schemas/CampaignStatus'
        subject:
          type: string
        recipientCount:
          type: integer
          minimum: 0
        scheduledAt:
          type: [string, 'null']
          format: date-time
        sentAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
    CampaignSummaryResponse:
      type: object
      required: [id, name, status, recipientCount, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          $ref: '#/components/schemas/CampaignStatus'
        recipientCount:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
...
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - code
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        path:
          type: string
        traceId:
          type: [string, 'null']
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'